requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pypdf
langchain-google-genai
langchain